Revert "feat: rename results to silva"
This reverts commit 02010ad870fb631553a07e61c3d1c764353dd068.
