+++ backend/openshift.deploy.yml
@@ -44,7 +44,7 @@ parameters:
   - name: DB_POOL_MAX_LIFETIME
     description: Maximum lifetime of a connection in the pool.
     value: "1800000"
-  - name: SILVA_ENV_OPENSEARCH
+  - name: RESULTS_ENV_OPENSEARCH
     description: Environment name for OpenSearch. # One of: development, test, production
     required: true
     value: development
@@ -121,8 +121,8 @@ objects:
                   value: ${DB_POOL_IDLE_TIMEOUT}
                 - name: DB_POOL_MAX_LIFETIME
                   value: ${DB_POOL_MAX_LIFETIME}
-                - name: SILVA_ENV_OPENSEARCH
-                  value: ${SILVA_ENV_OPENSEARCH}
+                - name: RESULTS_ENV_OPENSEARCH
+                  value: ${RESULTS_ENV_OPENSEARCH}
                 - name: DATABASE_HOST
                   value: ${DATABASE_HOST}
                 - name: DATABASE_PORT
