+++ common/openshift.init.yml
@@ -65,7 +65,7 @@ objects:
     metadata:
       name: allow-same-namespace
       labels:
-        template: nr-silva-backend-network-security-policy
+        template: nr-results-backend-network-security-policy
     spec:
       ingress:
         - from:
